Rename `ICanBoogie\Core\ConfigureEvent` as `ICanBoogie\Application\ConfigureEvent`
