+++ lib/Application/ConfigureEvent.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace ICanBoogie\Core;
+namespace ICanBoogie\Application;
 
 use ICanBoogie\Application;
 use ICanBoogie\Event;
 
 /**
- * Event class for the `ICanBoogie\Core::configure` event.
+ * Event class for the `ICanBoogie\Application::configure` event.
  *
  * The event is fired when the application is configured.
  *
